requirements.txt
arrow==1.3.0
beautifulsoup4==4.12.3
blinker==1.8.2
bokeh==3.6.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comtypes==1.4.7
contourpy==1.3.0
cssselect==1.2.0
easytrader==0.23.0
easyutils==0.1.7
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Logbook==1.7.0.post0
lxml==5.3.0
MarkupSafe==3.0.1
numpy==1.26.4
packaging==24.1
pandas==2.2.3
pillow==10.4.0
py-mini-racer==0.6.0
pycryptodome==3.21.0
PyMySQL==1.1.1
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywinauto==0.6.6
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.35
tornado==6.4.1
tqdm==4.66.5
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
xyzservices==2024.9.0
supervisor==4.2.5
mysqlclient==2.2.4
